refactor: move auth classes to appropriate packages
