+++ app/src/main/java/com/example/ecommerceapp/MainViewModel.kt
@@ -2,14 +2,14 @@ package com.example.ecommerceapp
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.ecommerceapp.data.auth.AuthManager
+import com.example.ecommerceapp.data.repository.AuthRepository
 import com.example.ecommerceapp.model.UserSession
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
 
 class MainViewModel(
-    authManager: AuthManager
+    authManager: AuthRepository
 ) : ViewModel() {
     val activeSession: StateFlow<UserSession?> = authManager.getActiveSession()
         .stateIn(
