+++ app/src/main/java/com/example/ecommerceapp/data/auth/AuthResult.kt
@@ -1,5 +1,7 @@
 package com.example.ecommerceapp.data.auth
 
+import com.example.ecommerceapp.model.UserApiResponse
+
 
 sealed class AuthResult {
     data class Success(val user: UserApiResponse? = null) : AuthResult()
