+++ app/src/main/java/com/example/ecommerceapp/data/core/APIService.kt
@@ -1,6 +1,6 @@
 package com.example.ecommerceapp.data.core
 
-import com.example.ecommerceapp.data.auth.AuthService
+import com.example.ecommerceapp.data.auth.RemoteAuthService
 import com.example.ecommerceapp.data.cart.CartApi
 import com.example.ecommerceapp.data.notification.RemoteNotificationApiDataSource
 import com.example.ecommerceapp.data.product.ProductApi
@@ -17,7 +17,7 @@ object APIService {
     val cartApi: CartApi = client.create(CartApi::class.java)
     val productApi: ProductApi = client.create(ProductApi::class.java)
     val categoryApi: CategoryApi = client.create(CategoryApi::class.java)
-    val authService: AuthService = client.create(AuthService::class.java)
+    val remoteAuthApi: RemoteAuthService = client.create(RemoteAuthService::class.java)
 
     val notificationApi: RemoteNotificationApiDataSource = client.create(
         RemoteNotificationApiDataSource::class.java
