+++ app/src/main/java/com/example/ecommerceapp/data/repository/AuthRepository.kt
@@ -1,17 +1,22 @@
-package com.example.ecommerceapp.data.auth
+package com.example.ecommerceapp.data.repository
 
+import com.example.ecommerceapp.data.auth.AuthResult
+import com.example.ecommerceapp.data.auth.RemoteAuthService
+import com.example.ecommerceapp.data.auth.UserSessionDao
+import com.example.ecommerceapp.model.LoginRequest
+import com.example.ecommerceapp.model.SignUpRequest
 import com.example.ecommerceapp.model.UserSession
 import kotlinx.coroutines.flow.Flow
 import java.io.IOException
 
-class AuthManager(
-    private val authService: AuthService,
+class AuthRepository(
+    private val remoteAuthService: RemoteAuthService,
     private val userSessionDao: UserSessionDao
 ) {
     suspend fun register(name: String, email: String, password: String): AuthResult {
         return try {
             val request = SignUpRequest(fullName = name, email = email, password = password)
-            val response = authService.signup(request)
+            val response = remoteAuthService.signup(request)
 
             if (response.isSuccessful) {
                 AuthResult.Success()
@@ -30,7 +35,7 @@ class AuthManager(
     suspend fun login(email: String, password: String): AuthResult {
         return try {
             val request = LoginRequest(email = email, password = password)
-            val response = authService.login(request)
+            val response = remoteAuthService.login(request)
 
             if (response.isSuccessful && response.body() != null) {
                 val userFromApi = response.body()!!.user
