--- app/src/main/java/com/example/ecommerceapp/screens/auth/SignUpViewModelFactory.kt
@@ -1,17 +1,0 @@
-package com.example.ecommerceapp.screens.auth
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.example.ecommerceapp.data.auth.AuthManager
-class SignUpViewModelFactory(
-    private val authManager: AuthManager
-) : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(SignUpViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return SignUpViewModel(authManager) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
